refactoring + circular imports fixed
