+++ src/core.jl
@@ -6,8 +6,6 @@ using SoleModels: bestguess
 using Parameters
 using FillArrays
 using StatsBase
-using ModalDecisionLists.LossFunctions: laplace_accuracy
-using ModalDecisionLists.LossFunctions: significance_test
 
 const SatMask = BitVector
 
@@ -190,215 +188,3 @@ Implementations should define at least:
 abstract type AbstractGenerator end
 
 ############################################################################################
-
-
-include("searchmethods/beamsearch.jl")
-include("searchmethods/randsearch.jl")
-include("searchmethods/atom-generator.jl")
-include("searchmethods/random-generator.jl")
-
-
-"""
-maptointeger(y::AbstractVector{<:CLabel})
-
-Map a categorical label vector to 1-based integer codes and return the ordered unique labels.
-
-Arguments
-- y: AbstractVector whose element type is a subtype of CLabel. Labels are compared using `==`.
-
-Returns
-- integer_y::Vector{UInt32}: a vector of length `length(y)` containing 1-based integer codes. Each distinct label v in `values` is assigned the code `i` where `values[i] == v`.
-- values::Vector{eltype(y)}: the unique labels appearing in `y`, in the order of their first occurrence (as produced by `unique(y)`).
-
-Notes
-- The mapping is stable with respect to the first occurrence order of labels in `y`.
-- The current implementation determines codes by comparing each label against the list of unique values (cost roughly O(n*m) where m = number of unique labels). For very large numbers of distinct labels a dictionary-based approach may be more efficient.
-
-Example
-```
-julia> y = ["red","blue","red"]
-julia> codes, vals = maptointeger(y)
- codes == UInt32[1,2,1]
- vals  == ["red","blue"]
-```
-"""
-function maptointeger(y::AbstractVector{<:CLabel})
-
-    # ordered values
-    values = unique(y)
-    integer_y = zeros(UInt32, length(y))
-
-    for (i, v) in enumerate(values)
-        integer_y[y.==v] .= i
-    end
-    return integer_y, values
-end
-
-
-"""
-    get_binary_labels_distribution(y::AbstractVector{<:Integer}, w::AbstractVector, target_class::Union{Integer,Nothing} = nothing) -> Tuple
-
-Compute the weighted distribution of binary labels for a given target class.
-
-This function calculates the sum of weights for positive and negative instances, where positive 
-instances are those whose labels match the target class, and negative instances are all others.
-If target class is nothing, the tuple (0,0) is returned
-
-# Arguments
-- `y::AbstractVector{<:Integer}`: Vector of class labels.
-- `w::AbstractVector`: Vector of weights corresponding to each instance.
-- `target_class::Union{Integer,Nothing}`: The class label to treat as positive. If `nothing`, 
-  defaults to binary classification with the first unique label as positive. Default: `nothing`.
-
-# Returns
-- `Tuple`: A tuple `(p, n)` where:
-  - `p`: Sum of weights for positive instances (labels equal to `target_class`).
-  - `n`: Sum of weights for negative instances (labels not equal to `target_class`).
-
-"""
-function get_binary_labels_distribution(
-    y::AbstractVector{<:Integer}, 
-    w::AbstractVector,
-    target_class::Union{Integer,Nothing} = nothing
-)
-    if isnothing(target_class)
-        return 0,0
-    end
-
-    pos_mask = (y .== target_class)
-    neg_mask = (y .!= target_class)
-    p = sum(w[pos_mask])
-    n = sum(w[neg_mask])
-    return p, n
-end
-
-
-"""
-    TODO sortantecedents....
-
-
-Takes an *antecedents*, each decorated by a SatMask indicating his coverage bitmask.
-Each antecedent is evaluated on his covered y using the provided *loss_function* function.
-Then the permutation of the bests *beam_width* sorted antecedent is returned with the lossfnctn
-value of the best one.
-
-See also
-[`entropy`](@ref).
-"""
-function sortantecedents(
-    antecedents::AbstractVector{Antecedent},
-    y::AbstractVector{<:CLabel},
-    w::AbstractVector,
-    beam_width::Integer,
-    loss_function::SymmetricLoss,
-    min_rule_coverage::Integer,
-    max_infogain_ratio::Union{Real,Nothing},
-    significance_alpha::Union{Real,Nothing};
-    kwargs...
-)::Tuple{AbstractVector,<:Real}     # TODO: Dispatching in base a tipo loss_function
-    isempty(antecedents) && return [], Inf
-
-    # If 'min_rule_coverage' is defined, this filters out from antecedents any antecedent whose covmasks covers less than 'min_rule_coverage' samples 
-    if min_rule_coverage > 1
-        validindices = findall(ant -> count(ant.covmask) >= min_rule_coverage, antecedents)
-        isempty(validindices) && return [], Inf
-        antecedents = antecedents[validindices]
-    end
-
-    indices = eachindex(antecedents)
-
-    # loss function values for each antecedent
-    antslossfnctn = map(a ->  loss_function(y, w; antecedent=a, kwargs...) , antecedents)
-
-    if !isnothing(max_infogain_ratio)
-        # every rule whose loss is < const. * loss of ⊤ over dataset is to be removed, this makes the actual sorting at the end faster
-        minloss = (1-max_infogain_ratio) * loss_function(y, w; kwargs...)
-
-        # Keep only the indices corresponding antecedents whose loss is ≥ min_loss
-        indices = [ind for (ind, loss) in enumerate(antslossfnctn) if loss ≥ minloss]
-
-        isempty(indices) && return [], Inf
-    end
-
-    # Extract the indices (with respect to antslossfnctn) of the 'beam_width' best antecedents (with lowest loss)
-    valid_indices = partialsortperm(antslossfnctn[indices], 1:min(beam_width, length(indices)))
-
-    newstar_perm = indices[valid_indices]  # convert indices to those relative to the parameter antecedents
-    newstar = antecedents[newstar_perm]    # extract best antecedents and corresponding loss functions
-    bestantecedent_lossfnctn = antslossfnctn[newstar_perm[1]]
-
-    return newstar, bestantecedent_lossfnctn
-end
-
-
-
-function sortantecedents(
-    antecedents::AbstractVector{Antecedent},
-    y::AbstractVector{<:CLabel},
-    w::AbstractVector,
-    beam_width::Integer,
-    loss_function::AsymmetricLoss,
-    min_rule_coverage::Integer,
-    max_infogain_ratio::Union{Real,Nothing},
-    significance_alpha::Union{Real,Nothing};
-    target_class::Union{Integer, Nothing},
-    kwargs...
-)::Tuple{AbstractVector,<:Real}     # TODO: Dispatching in base a tipo loss_function
-    isempty(antecedents) && return [], Inf
-
-    # If 'min_rule_coverage' is defined, this filters out from antecedents any antecedent whose covmasks covers less than 'min_rule_coverage' samples 
-    if min_rule_coverage > 1
-        validindices = findall(ant -> count(ant.covmask) >= min_rule_coverage, antecedents)
-        isempty(validindices) && return [], Inf
-        antecedents = antecedents[validindices]
-    end
-
-    indices = eachindex(antecedents)
-
-    # loss function values for each antecedent
-    antslossfnctn = map(a ->  loss_function(y, w, target_class; antecedent=a, kwargs...) , antecedents)
-
-    # if !isnothing(max_infogain_ratio)
-    #     # every rule whose loss is < const. * loss of ⊤ over dataset is to be removed, this makes the actual sorting at the end faster
-    #     minloss = (1-max_infogain_ratio) * loss_function(y, w, target_class; kwargs...)
-
-    #     # Keep only the indices corresponding antecedents whose loss is ≥ min_loss
-    #     indices = [ind for (ind, loss) in enumerate(antslossfnctn) if loss ≥ minloss]
-
-    #     isempty(indices) && return [], Inf
-    # end
-
-    # Extract the indices (with respect to antslossfnctn) of the 'beam_width' best antecedents (with lowest loss)
-    valid_indices = partialsortperm(antslossfnctn[indices], 1:min(beam_width, length(indices)))
-
-    newstar_perm = indices[valid_indices]  # convert indices to those relative to the parameter antecedents
-    newstar = antecedents[newstar_perm]    # extract best antecedents and corresponding loss functions
-    bestantecedent_lossfnctn = antslossfnctn[newstar_perm[1]]
-
-    return newstar, bestantecedent_lossfnctn
-end
-
-
-
-
-
-############################################################################################
-############ Utils #########################################################################
-############################################################################################
-
-function preprocess_inputdata(
-    X::AbstractDataFrame,
-    y;
-    remove_duplicate_rows=false
-)
-    if remove_duplicate_rows
-        allunique(X) && return (X, y)
-        nonunique_ind = nonunique(X)
-        Xy = hcat(X[findall((!).(nonunique_ind)), :],
-            y[findall((!).(nonunique_ind))]
-        ) |> dropmissing
-    else
-        Xy = hcat(X[:, :], y[:]) |> dropmissing
-    end
-    return Xy[:, 1:(end-1)], Xy[:, end]
-end
