+++ src/loss-functions.jl
@@ -6,48 +6,8 @@ using SoleModels
 using FillArrays
 using StatsBase
 using Distributions
-
-"""
-    count_labels_distribution(y::AbstractVector{<:UInt32}, 
-        nlabels::Integer, 
-        w=default_weights(length(y))
-    )::Tuple{AbstractVector{<:Real}, Integer}
-
-
-Compute the weighted distribution of labels in a vector.
-
-# Arguments
-- `y::AbstractVector{<:UInt32}`: A vector of label indices.
-- `nlabels::Integer`: The total number of distinct labels.
-- `w`: Optional weight vector for each sample. Defaults to uniform weights if not provided.
-
-# Returns
-A tuple containing:
-- `dist::AbstractVector{<:Real}`: A vector where `dist[i]` is the weighted count of occurrences of label `i`.
-- `y_offset::Integer`: The offset applied to normalize labels so the minimum label value becomes 1.
-
-# Details
-The function internally adjusts the label indices by offsetting them so that the minimum label value becomes 1.
-This adjustment is necessary to properly index into the distribution vector. The offset value is returned
-to allow for later denormalization if needed.
-"""
-function count_labels_distribution( y::AbstractVector{<:UInt32},
-    n::Integer,
-    w=default_weights(length(y))
-)::Tuple{AbstractVector{<:Real}, Integer}
-
-    y_min = convert(Int64, minimum(y)) # this is necesssary, otherwise -y_min underflows when calculating y_offset
-    y_offset = -y_min + 1
-
-    # y_adj is just y normalized with a constant offset so that the smallest value is +1 
-    y_adj = convert.(Int64, y) .+ y_offset
-
-    # dist[i] is the number of occurrences of the label i in y_adj
-    dist = counts(y_adj, n, Weights(w))
-
-    return dist, y_offset   
-end # TODO: Spostare sta roba da qualche altra parte
-
+using ..ModalDecisionLists: Antecedent
+using ..Metrics
 
 ############################################################################################
 ############################# New Loss Functions ###########################################
@@ -62,6 +22,11 @@ abstract type SymmetricLoss <: AbstractLossFunction end
 abstract type AsymmetricLoss <: AbstractLossFunction end
 
 
+export AbstractLossFunction
+export SymmetricLoss
+export AsymmetricLoss
+
+
 function calculate_loss(::AbstractLossFunction; kwargs...)
     error("calculate_loss can only be called with a non-abstract loss function type")
 end
@@ -74,18 +39,13 @@ end
 # Implementa le loss come callable structs
 struct GiniImpurity <: SymmetricLoss end
 
+export GiniImpurity
+
 function (::GiniImpurity)(
     y::AbstractVector{<:Integer},
     w::AbstractVector = default_weights(length(y))
 )
-    isempty(y) && return Inf
-    
-    dist = w isa Ones ? counts(y) : counts(y, Weights(w))
-    filter!(!iszero, dist)
-    length(dist) == 1 && return 0.0
-    
-    p = dist ./ sum(dist)
-    return 1 - sum(abs2, p)  # abs2 è più efficiente di .^2
+    return gini_impurity(y, w)                      # lower level implementations of metrics in metrics.jl
 end
 
 #######################
@@ -98,26 +58,23 @@ end
 # ENTROPY
 struct Entropy <: SymmetricLoss end
 
+export Entropy
+
+
 function (::Entropy)(
     y::AbstractVector{<:Integer},
     w::AbstractVector=default_weights(length(y));
     kwargs...
 )
-    isempty(y) && return Inf
-
-    distribution = (w isa Ones ? counts(y) : counts(y, Weights(w)))
-    distribution = distribution[distribution .!= 0]
-
-    length(distribution) == 1 && return 0.0
-
-    prob = distribution ./ sum(distribution)
-    e = -sum(prob .* log2.(prob))
-    return e
+    return entropy(y, w; kwargs...)                 # lower level implementations of metrics in metrics.jl
 end
 
 
 # LAPLACE METRIC
 struct LaplaceMetric <: SymmetricLoss end
+
+export LaplaceMetric
+
 
 function (::LaplaceMetric)(
     y::AbstractVector{<:UInt32},
@@ -125,17 +82,7 @@ function (::LaplaceMetric)(
     nlabels::Integer,
     kwargs...
 )
-    isempty(y) && return 0.0
-
-    @assert length(w) == length(y) "weights and labels must have the same length"
-
-    dist, y_offset = count_labels_distribution(y, nlabels, w)
-
-    # number of effective labels (k) and number of matches for target class (target)
-    k, target = nothing, nothing
-    k, target = nlabels, maximum(dist)
-
-    return 1 - ((target + 1) / (sum(dist) + k))
+    return laplace_metric(y, w; nlabels, kwargs...)         # lower level implementations of metrics in metrics.jl
 end
 
 #####################################################
@@ -145,6 +92,9 @@ end
 
 # FOIL GAIN
 struct FOILGain <: AsymmetricLoss end
+
+export FOILGain
+
 
 function (::FOILGain)(
     y::AbstractVector{<:UInt32},
@@ -186,12 +136,15 @@ end
 # LAPLACE ACCURACY
 struct LaplaceAccuracy <: AsymmetricLoss end    
 
+export LaplaceAccuracy
+
+
 function (::LaplaceAccuracy)(
     y::AbstractVector{<:UInt32},
     w::AbstractVector,
     target_class::Integer;
     antecedent::Antecedent = nothing,
-    prev_antecedent::Antecedent = nothing,
+    prev_antecedent::Union{Antecedent, Nothing} = nothing,
     nlabels::Integer,
     kwargs...
 )
@@ -206,121 +159,4 @@ function (::LaplaceAccuracy)(
     return 1 - (tp + 1) / (tp + fp + 2)
 end
 
-
-
-
-############################################################################################
-############################# Old Loss Functions ###########################################
-
-
-"""
-    laplace_accuracy(y::AbstractVector{<:UInt32}, w::AbstractVector=default_weights(length(y)); nlabels::Integer, target_class::Union{Integer,Nothing}=nothing, kwargs...) -> Float64
-
-Computes a Laplace-accuracy based loss metric to estimate the coverage level of a rule.
-A lower returned value indicates better coverage of target_class on y.
-
-# Arguments
-- `y::AbstractVector{<:UInt32}`: Vector of class labels (0-based or 1-based indexing).
-- `w::AbstractVector`: Vector of weights for each sample. Defaults to uniform weights.
-- `nlabels::Integer`: Total number of distinct labels in the dataset.
-- `target_class::Union{Integer,Nothing}`: Specific class to evaluate. If `nothing`, uses the class with maximum count.
-- `kwargs...`: Additional keyword arguments (unused).
-
-# Returns
-- `Float64`: A Laplace-smoothed accuracy value in the range [0, 1], where higher values indicate better accuracy.
-
-# Details
-The Laplace smoothing formula is: `1 - ((target + 1) / (sum(dist) + k))`
-where:
-- `target`: Count of samples matching the target class
-- `sum(dist)`: Total count across all classes
-- `k`: Number of effective labels (2 if `target_class` is specified, otherwise `nlabels`)
-
-# Throws
-- `AssertionError`: If the lengths of `w` and `y` do not match.
-"""
-function laplace_accuracy(
-    y::AbstractVector{<:UInt32},
-    w::AbstractVector=default_weights(length(y));
-    nlabels::Integer,
-    target_class::Union{Integer,Nothing} = nothing,
-    kwargs...
-)
-    isempty(y) && return 0.0
-
-    @assert length(w) == length(y) "weights and labels must have the same length"
-
-    dist, y_offset = count_labels_distribution(y, nlabels, w)
-
-    # number of effective labels (k) and number of matches for target class (target)
-    k, target = nothing, nothing
-    if !isnothing(target_class)
-        # Converti target_class se era 0-based
-        target_class_adj = target_class .+ y_offset
-
-        # println("Number of pos elements in y: $(length(findall(label -> label==1, y)))")
-        # println("Number of pos elements in y_adj: $(length(findall(label -> label==target_class_adj, y_adj)))")
-
-        # println("y: $y")
-        # println("y adj: $y_adj")
-
-        # println("target class adj: $target_class_adj")
-        # println("dist: $dist")
-        k, target = 2, dist[Int(target_class_adj)]
-    else
-        k, target = nlabels, maximum(dist)
-    end
-
-    return 1 - ((target + 1) / (sum(dist) + k))
 end
-
-function entropy(
-    y::AbstractVector{<:Integer},
-    w::AbstractVector=default_weights(length(y));
-    kwargs...
-)
-    isempty(y) && return Inf
-
-    distribution = (w isa Ones ? counts(y) : counts(y, Weights(w)))
-    distribution = distribution[distribution .!= 0]
-
-    length(distribution) == 1 && return 0.0
-
-    prob = distribution ./ sum(distribution)
-    e = -sum(prob .* log2.(prob))
-    return e
-end
-
-# ############################################################################################
-# ############################# Significance Test ############################################
-
-function significance_test(
-    ycurr::AbstractVector{<:Integer},
-    yprev::AbstractVector{<:Integer},
-    alpha::Real;
-    target_class::Union{Integer,Nothing} = nothing,
-    nlabels::Integer,
-    kwargs...
-)
-    currdist = counts(ycurr, nlabels)
-    prevdist = counts(yprev, nlabels)
-    if !isnothing(target_class)
-        x = Vector{Real}([currdist[tc], sum(currdist) - currdist[tc]])
-        y = Vector{Real}([prevdist[tc], sum(prevdist) - prevdist[tc]])
-    else
-        x = Vector{Real}(currdist)
-        y = Vector{Real}(prevdist)
-    end
-    lrs = begin
-        x[x .== 0] .= 1e-5
-        y[y .== 0] .= 1e-5
-        y = y * (sum(x)/sum(y))
-        # Likelihood Ratio Statistic
-        sum(x .* log.(x ./ y)) * 2
-    end
-    # Degrees of freedom
-    df = length(currdist) - 1
-    return ( lrs > 0 ) & (ccdf(Chisq(df), lrs) <= alpha)
-end
-
-end # module
