+++ src/search.jl
@@ -1,0 +1,122 @@
+using SoleBase: CLabel
+using SoleData: AbstractLogiset, PropositionalLogiset
+using SoleModels: bestguess
+using Parameters
+using FillArrays
+using StatsBase
+
+
+
+include("searchmethods/beamsearch.jl")
+include("searchmethods/randsearch.jl")
+include("searchmethods/atom-generator.jl")
+include("searchmethods/random-generator.jl")
+
+
+
+"""
+    TODO sortantecedents....
+
+
+Takes an *antecedents*, each decorated by a SatMask indicating his coverage bitmask.
+Each antecedent is evaluated on his covered y using the provided *loss_function* function.
+Then the permutation of the bests *beam_width* sorted antecedent is returned with the lossfnctn
+value of the best one.
+
+See also
+[`entropy`](@ref).
+"""
+function sortantecedents(
+    antecedents::AbstractVector{Antecedent},
+    y::AbstractVector{<:CLabel},
+    w::AbstractVector,
+    beam_width::Integer,
+    loss_function::SymmetricLoss,
+    min_rule_coverage::Integer,
+    max_infogain_ratio::Union{Real,Nothing},
+    significance_alpha::Union{Real,Nothing};
+    kwargs...
+)::Tuple{AbstractVector,<:Real}     # TODO: Dispatching in base a tipo loss_function
+    isempty(antecedents) && return [], Inf
+
+    # If 'min_rule_coverage' is defined, this filters out from antecedents any antecedent whose covmasks covers less than 'min_rule_coverage' samples 
+    if min_rule_coverage > 1
+        validindices = findall(ant -> count(ant.covmask) >= min_rule_coverage, antecedents)
+        isempty(validindices) && return [], Inf
+        antecedents = antecedents[validindices]
+    end
+
+    indices = eachindex(antecedents)
+
+    # loss function values for each antecedent
+    antslossfnctn = map(a ->  loss_function(y, w; antecedent=a, kwargs...) , antecedents)
+
+    if !isnothing(max_infogain_ratio)
+        # every rule whose loss is < const. * loss of ⊤ over dataset is to be removed, this makes the actual sorting at the end faster
+        minloss = (1-max_infogain_ratio) * loss_function(y, w; kwargs...)
+
+        # Keep only the indices corresponding antecedents whose loss is ≥ min_loss
+        indices = [ind for (ind, loss) in enumerate(antslossfnctn) if loss ≥ minloss]
+
+        isempty(indices) && return [], Inf
+    end
+
+    # Extract the indices (with respect to antslossfnctn) of the 'beam_width' best antecedents (with lowest loss)
+    valid_indices = partialsortperm(antslossfnctn[indices], 1:min(beam_width, length(indices)))
+
+    newstar_perm = indices[valid_indices]  # convert indices to those relative to the parameter antecedents
+    newstar = antecedents[newstar_perm]    # extract best antecedents and corresponding loss functions
+    bestantecedent_lossfnctn = antslossfnctn[newstar_perm[1]]
+
+    return newstar, bestantecedent_lossfnctn
+end
+
+
+
+
+
+function sortantecedents(
+    antecedents::AbstractVector{Antecedent},
+    y::AbstractVector{<:CLabel},
+    w::AbstractVector,
+    beam_width::Integer,
+    loss_function::AsymmetricLoss,
+    min_rule_coverage::Integer,
+    max_infogain_ratio::Union{Real,Nothing},
+    significance_alpha::Union{Real,Nothing};
+    target_class::Union{Integer, Nothing},
+    kwargs...
+)::Tuple{AbstractVector,<:Real}     # TODO: Dispatching in base a tipo loss_function
+    isempty(antecedents) && return [], Inf
+
+    # If 'min_rule_coverage' is defined, this filters out from antecedents any antecedent whose covmasks covers less than 'min_rule_coverage' samples 
+    if min_rule_coverage > 1
+        validindices = findall(ant -> count(ant.covmask) >= min_rule_coverage, antecedents)
+        isempty(validindices) && return [], Inf
+        antecedents = antecedents[validindices]
+    end
+
+    indices = eachindex(antecedents)
+
+    # loss function values for each antecedent
+    antslossfnctn = map(a ->  loss_function(y, w, target_class; antecedent=a, kwargs...) , antecedents)
+
+    # if !isnothing(max_infogain_ratio)
+    #     # every rule whose loss is < const. * loss of ⊤ over dataset is to be removed, this makes the actual sorting at the end faster
+    #     minloss = (1-max_infogain_ratio) * loss_function(y, w, target_class; kwargs...)
+
+    #     # Keep only the indices corresponding antecedents whose loss is ≥ min_loss
+    #     indices = [ind for (ind, loss) in enumerate(antslossfnctn) if loss ≥ minloss]
+
+    #     isempty(indices) && return [], Inf
+    # end
+
+    # Extract the indices (with respect to antslossfnctn) of the 'beam_width' best antecedents (with lowest loss)
+    valid_indices = partialsortperm(antslossfnctn[indices], 1:min(beam_width, length(indices)))
+
+    newstar_perm = indices[valid_indices]  # convert indices to those relative to the parameter antecedents
+    newstar = antecedents[newstar_perm]    # extract best antecedents and corresponding loss functions
+    bestantecedent_lossfnctn = antslossfnctn[newstar_perm[1]]
+
+    return newstar, bestantecedent_lossfnctn
+end
