+++ test/benchmark.jl
@@ -12,7 +12,7 @@ using ModalDecisionLists
 using ModalDecisionLists: BaseCN2, MLJInterface
 using ModalDecisionLists: preprocess_inputdata
 using CategoricalArrays: CategoricalValue, CategoricalArray
-using ModalDecisionLists. LossFunctions: laplace_accuracy
+using ModalDecisionLists: laplace_accuracy
 using BenchmarkTools
